Templating General pattern
non functional code
